+++ usr/src/uts/sun4u/io/wrsm/wrsm_nr.c
@@ -2,9 +2,8 @@
  * CDDL HEADER START
  *
  * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
+ * Common Development and Distribution License (the "License").
+ * You may not use this file except in compliance with the License.
  *
  * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
  * or http://www.opensolaris.org/os/licensing.
@@ -20,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2006 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -6991,7 +6990,8 @@ update_wrsm_route_kstat(kstat_t *ksp, int rw)
 				    "current stripe total for passthrough %d",
 				    rte_named->num_stripes.value.ui32));
 			} else {
-				rte_named->nodeid[i][j].value.ui32 = -1;
+				rte_named->nodeid[i][j].value.ui32 =
+				    (uint32_t)-1;
 			}
 
 			rte_named->gnid[i][j].value.ui32 =
